+++ qemu-0.12.5/module/history.c
@@ -11,8 +11,8 @@ struct history_entry {
 struct history_queue {
     struct history_entry load_entry[MAX_LOAD_QUEUE_SIZE];
     struct history_entry store_entry[MAX_STORE_QUEUE_SIZE];
-    uint32_t load_tail;
-    uint32_t store_tail;
+    uint64_t load_tail;
+    uint64_t store_tail;
 };
 
 #define MAX_HASH_NUM 1024
@@ -44,9 +44,9 @@ static inline void module_history_load_record(struct trace_content *content)
 {
 #ifdef MOD_HISTORY
     uint8_t tid;
-    uint32_t address;
+    uint64_t address;
     struct history_queue *temp_queue;
-    uint32_t tail;
+    uint64_t tail;
     struct history_entry *temp_entry;
 
     tid = content->tid;
@@ -70,9 +70,9 @@ static inline void module_history_store_record(struct trace_content *content)
 {
 #ifdef MOD_HISTORY
     uint8_t tid;
-    uint32_t address;
+    uint64_t address;
     struct history_queue *temp_queue;
-    uint32_t tail;
+    uint64_t tail;
     struct history_entry *temp_entry;
 
     tid = content->tid;
@@ -98,10 +98,10 @@ void module_match_with_load(struct trace_content *content, uint8_t other_tid)
 {
 #ifdef MOD_MATCH
     uint8_t tid;
-    uint32_t address, other_address;
-    uint32_t index, other_index, last_index;
+    uint64_t address, other_address;
+    uint64_t index, other_index, last_index;
     struct history_queue *temp_queue;
-    uint32_t head, tail;
+    uint64_t head, tail;
     struct history_entry *temp_entry;
 
     tid = content->tid;
@@ -155,10 +155,10 @@ void module_match_with_store(struct trace_content *content, uint8_t other_tid)
 {
 #ifdef MOD_MATCH
     uint8_t tid;
-    uint32_t address, other_address;
-    uint32_t index, other_index, last_index;
+    uint64_t address, other_address;
+    uint64_t index, other_index, last_index;
     struct history_queue *temp_queue;
-    uint32_t head, tail;
+    uint64_t head, tail;
     struct history_entry *temp_entry;
 
     tid = content->tid;
